prepare 3.1.1.0.11 release of ruby-maven
the problem was that version of the bundled telsa-polyglot-ruby
needs to be on maven-central to load on the fly the missing
jruby bits.

this is build of 0.1.2-SNAPSHOT with get relabled with 0.1.1 (bad thing
but repo needs a new home once maven-3.6.2 is released and then rectify this
version mess)

fixes https://github.com/mkristian/jbundler/issues/58

Sponsored by Lookout Inc.